workflows/part2_version2/requirements.txt
catboost==1.2.5
lightgbm==4.4.0
openpyxl==3.1.4
pubchempy==1.0.4
pyyaml==6.0.1 
rdkit==2024.3.1
scikit-learn==1.2.2
scipy==1.11.1
seaborn==0.12.2
shap==0.46.0
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
xgboost==2.1.0
# Version checked by AK and LT 7th Oct, 2024